+++ app/org/fairs/[fairId]/sales/page.tsx
@@ -40,6 +40,7 @@ export default async function OrgSalesFeedPage({
         initialSales={initial.sales}
         initialTotalUnits={initial.totalUnits}
         initialTotalRevenue={initial.totalRevenue}
+        initialCashRevenue={initial.cashRevenue}
         initialPayout={initial.payout}
         initialPayoutIsFinal={initial.payoutIsFinal}
         initialMissingInventoryCost={initial.missingInventoryCost}